Blueprints: Throw on failure
This commit ensures the Blueprint steps will throw an error when they
fail. For now, the error is merely reported in the console. In the future,
there may be a nice UI for this.

Detecting a runPHP failure required adjusting the php_wasm.c code to return
the correct exit code. The previous method of inferring the exit code was
replaced by simply returning `EG(exit_status)` which is populated by the
zend engine.

Furthermore, this PR ships additional unit tests for some Blueprint
steps to ensure they indeed throw as expected.

 ## Testing instructions

This PR comes with test coverage so confirm all the tests pass in CI.

Also, apply this PR locally and visit the following URL:

http://localhost:5400/website-server/#{%22steps%22:[{%22step%22:%22runPHP%22,%20%22code%22:%22%3C?php%20no_such_fn();%22}]}

It should reveal useful error information in the console.