PHP: Pass request body as Uint8Array
Removes the custom file upload handler and rely on PHP body parsing
to populate the $_FILES array. Instead of encoding the body bytes as
a string, parsing that string, and re-encoding it as bytes, we keep
the body in a binary form and pass it directly to PHP HEAP memory.

Closes https://github.com/WordPress/wordpress-playground/issues/997
Closes https://github.com/WordPress/wordpress-playground/issues/1006
Closes https://github.com/WordPress/wordpress-playground/issues/914

 ## Testing instructions

Confirm the CI checks pass (it will take a few iterations to get them right I'm sure :D)