32bit integer workaround (#1014)
This is a temporary workaround to hide the 32bit integer warnings that
appear when using various time related function, such as strtotime() and
mktime().

## What problem is it solving?
Hiding the warning messages from being displayed on the page.

## How is the problem addressed?
Adding an error handler to ignore warning if it matches the message "fit
in a PHP integer"

## Testing Instructions
add ?year=123 to any playground URL (e.g. /?year=123). Normally it would
display `Warning: mktime(): Epoch doesn't fit in a PHP integer in
/wordpress/wp-includes/class-wp-date-query.php on line 3`