PHP: Do not prepend a whitespace when encoding body as multipart form data
When Playground request handler is called with a key: value object as
its body, that data is encoded as a multipart string:

```ts
await playground.request({
	url: `/wp-admin/import.php`,
	method: 'POST',
	body: { import: file },
});
```

However, the `encodeAsMultipart()` function prepended one newline too
many at the beginning of each multipart section.

Most of Playground do not use the `key: value` notation and just pass
the POST data as bytes, which is why the normal operating functions were
preserved.

However, the importFile step relied on the `key: value` notation.
Importing WXR involves submitting a form with a nonce. However, the
nonce contained an extra whitespace and was rejected by WordPress. This
PR ensures that extra whitespace isn't there.

 ## Testing instructions

* Ensure the CI tests pass