Website: Remove the colorful disco background
Due to an Edge bug on MacOS, the Playground viewport sometimes starts
blinking and the underlying background comes to the top at a rapid pace.
This is bad for people with visual sensitivity. Since we cannot fix
Edge, let's just remove the colorful background. It is not that relevant
ever since Playground uses the fullscreen mode by default.