Polyfill dns_check_record(), dns_get_record(), and dns_get_mx()
Closes https://github.com/WordPress/wordpress-playground/issues/1042 by
polyfilling the dns_* functions that are missing in the WebAssembly
build of PHP. In the PHP source code, these functions are included
conditionally when the HAVE_DNS_SEARCH_FUNC constant is true, which it
isn't in the Emscripten build.

The polyfills return either `false` or an empty array and do not
actually perform any DNS checks.

In Node.js, we could support actual dns checks. Let's leave that for a
subsequent iteration. In the browser, there isn't much we can do.

 ## Remaining work

Rebuild the browser versions of PHP

 ## Testing instructions

Confirm the CI checks pass

CC @p-jackson @sejas @wojtekn