Service worker: Improve error reporting in non-secure contexts
Playground errors in some non-secure contexts suggest service workers are not
supported in the current browser. This is confusing and the user is only
left with a "Playground had a hiccup" error message that looks like a
bug.

This PR changes the technique of checking whether
the current context is secure from a simple `https:` verification to
consulting window.isSecureContext and then surfaces that information in
the UI.

 ## Testing instructions

1. Start a new Local site on a .local domain. Make sure to use http://,
   not https://
1. Start Playground dev server
1. Open the local site, add an <iframe
   src="http://localhost:5400/website-server" > in devtools
1. Confirm the error message is useful and talks about HTTPS and secure contexts