Provide PHP Bison 2.7 dep for non-ARM64 arch
The prebuilt bison 2.7 files are for arm64 architecture and do not work
for other architectures such as x86_64.  In these cases, we can at least
build bison 2.7 from source.  We could be more efficient and build a
version that is cached across all PHP builds that need it, but let's get
the builds working again first.