Fix php bison dep for building on non-arm64 architectures (#1115)
This PR fixes an error building php-wasm for PHP 7.3 and below on
non-arm64 architectures. Those PHP versions depend on Bison 2.7. We have
a prebuilt version of Bison 2.7 for arm64, but that naturally doesn't
work on other CPU architectures. This PR builds Bison 2.7 as a stopgap.

On non-arm64 architectures, this extra build step runs once for each
supported PHP version up to PHP 7.3. This is not ideal but can be
improved later.

## Testing Instructions

### Before applying this patch

On an intel machine, run `npm run recompile:php:web` and observe there
is a build error.

### After applying this patch

On an intel machine, run `npm run recompile:php:web` and observe there
are no longer build errors.

On an arm64 machine, run the following to make sure there are no build
regressions:
`npm run recompile:php:web`