Web: Make the "Apply changes" button work in Playground settings form
Related to https://github.com/WordPress/wordpress-playground/issues/983

The "Apply changes" button in the "Customize Playground" modal doesn't
submit the form when the `playground` object isn't available. That may
happen when the remote.html API did not resolve yet, but also when the
Blueprint execution fails for any reason. For example, when the browser
storage contains outdated code that doesn't work anymore.

This PR makes the `playground` argument to reloadWithNewConfiguration optional.

 ## Follow up work

The downside is that a broken browser storage state means you may never
clear the OPFS storage as `resetVirtualOpfs()` is a function you call on
a `playground` instance.

 ## Testing instructions

This one is tricky. You need a broken Playground instance to test this.
The easiest way is to update the `0-playground.php` mu-plugin to trigger
a syntax error. Once you do that, confirm the modal cannot be submitted
without this PR and can with it.