Web: Make the "Apply changes" button work in Playground settings form (#1122)
The "Apply changes" button in the "Customize Playground" modal doesn't
submit the form when the `playground` object isn't available. That may
happen when the remote.html API did not resolve yet, but also when the
Blueprint execution fails for any reason. For example, when the browser
storage contains outdated code that doesn't work anymore.

This PR makes the `playground` argument to reloadWithNewConfiguration
optional.

It also adds a new `onClientConnected` argument to `startPlaygroundWeb`
to store the Playground client reference even if the Blueprint cannot be
executed.

 ## Testing instructions

This one is tricky. You need a broken Playground instance to test this.
The easiest way is to update the `0-playground.php` mu-plugin to trigger
a syntax error. Once you do that, confirm the modal cannot be submitted
without this PR and can with it.