CI: Deploy the playground.wordpress.net website every day at 9am
This fixes a regression introduced in an earlier commit.
build-website.yml used to be ran after a successful run of the "Refresh
WordPress assets" workflow. However, that Workflow now runs every 20
minutes, and such frequent deploys were undesirable so the completion
trigger was removed. That also disabled the daily auto-deploys.
Therefore, this commit restores them via a cron schedule.