Give brandonpayton permission to run Playground GH workflows (#1139)
This PR gives me permission to run Playground GitHub workflows. The
initial motivation for this is so I can manually run the needed
workflows immediately after the WordPress 6.5 release so
playground.wordpress.net is updated in a timely manner.

If this is not needed or desired at the moment, please feel free to
close. :)