PHP.run(): Throw JS exception on runtime error, remove throwOnError flag (#1137)
With this PR, `PHP.run()` throws an exception when `php.run()` is
unsuccessful. In other words, it makes `php.run({ throwOnError: true })`
the default behavior and removes the `throwOnError` flag.

## What problem is it solving?

With the introduction of the Playground logger, we want to collect all
errors in the logger and let the user decide what to do with the logged
errors. For this to work, we can't allow `throwOnError` to be `false`.

The early versions of Playground/PHP.wasm were meant to support this
behavior, but they couldn't because of the technical limitations – the
exit code captured from the PHP process did not reflect the truth about
the actual success/failure.

The `throwOnError` flag was introduced together with the correct
`exitCode` support. In the typical `php.run()` usage, you'll want to
check for errors. Asking the developers to always inspect the `exitCode`
is a mistake-prone pattern since it's easy to forget or get wrong.

With this PR, PHP.wasm does the error-checking work for the developer
and reliably reports any runtime crashes using JavaScript exceptions. It
also removes the non throwing– code path to reduce the number of
possible ways of using the library, reduce the number of uncaught errors
out in the wild, and ease the library maintenance.

### Before this PR:

```ts
const r = await php.run({
    code: `<?php noSuchFunction();`
});
console.log(r.errors); // Fatal Error: Undefined function...
```

### After this PR:

```ts
try {
    // This throws an exception:
    const r = await php.run({
        code: `<?php noSuchFunction();`
    });
} catch( e ) {
    console.log(e.output.stderr); // Fatal Error: Undefined function...
    console.log(e.output.stdout);
}
```

## How is the problem addressed?

By removing all mentions of `throwOnError` and making sure tests pass.

## Testing Instructions

- Confirm that all tests pass