Browser: Display PHP output when Fatal Error is trigerred (#1234)
## What does this PR do?

This PR restores displaying the PHP output when a PHP error is
encountered. It does that by attaching `response` to the error thrown by
BasePHP and augmenting the Comlink transfer handler to pass that
response between workers through postMessage.

 ## What problem does it solve?

Prior to be0e783, script output was still shown for Fatal Errors. For
example, when reproducing the memory-related errors under #1128, we used
to see something like the following where page output was shown along
with the Fatal Error that ended execution.

But now that we are no longer returning a PHPResponse when there is a
non-zero exit code, the content is not updated when running the same
script. The error message is printed to the console, but the partial
content is not visible to the user. Instead the previous content is left
in place as if the script had not run at all.

Closes https://github.com/WordPress/wordpress-playground/issues/1231

 ## Testing instructions

Ensure the E2E tests pass