Blueprints: Add ifAlreadyInstalled to installPlugin and installTheme steps (#1244)
Adds an `ifAlreadyInstalled?: 'overwrite' | 'skip' | 'error'` option to
installPlugin and installTheme steps. It defaults to `overwrite`.

Consider the following Blueprint:

```json
{
    "preferredVersions": {
        "php": "latest",
        "wp": "6.4"
    },
    "steps": [
        {
            "step": "installTheme",
            "themeZipFile": {
                "resource": "wordpress.org/themes",
                "slug": "twentytwentyfour"
            }
        }
    ]
}
```

Before this PR, it would result in an error. After this PR, the
installation just works. If the Blueprint author explicitly wants the
installation to fail, they can specify the `ifAlreadyInstalled` option:

```json
{
    "steps": [
        {
            "step": "installTheme",
            "themeZipFile": {
                "resource": "wordpress.org/themes",
                "slug": "twentytwentyfour"
            },
            "ifAlreadyInstalled": "skip" // or "error"
        }
    ]
}
```

## Motivation

Installing a plugin or theme over a currently installed one is a common
gotcha. Currently it results in an error and blocks the Blueprint
execution. This behavior is, however, often undesirable as it prevents
having a single Blueprint that installs a twentytwentyfour theme on
different versions of WordPress.

An addition of the `ifAlreadyInstalled` option puts the Blueprint author
in control and provides a sensible default behavior where the
installation will "just work" by replacing the already installed version
of the plugin or theme.

Closes https://github.com/WordPress/wordpress-playground/issues/1157
Related to https://github.com/adamziel/blueprints/pull/19

 ## Testing instructions

Confirm the unit tests pass

cc @bgrgicak @brandonpayton