Spawn new PHP instances for handling requests in PHPRequestHandler
The request handler needs to decide whether to serve a static asset or
run the PHP interpreter. For static assets it should just reuse the primary
PHP even if there's 50 concurrent requests to serve. However, for
dynamic PHP requests, it needs to grab an available interpreter.
Therefore, it cannot just accept PHP as an argument as serving requests
requires access to ProcessManager.