PHP: Add a cwd argument to hotSwapPHPRuntime() (#1304)
To support [Loopback
Request](https://github.com/WordPress/wordpress-playground/pull/1287),
we need to decouple BasePHP from the request handler. This PR removes
the implicit dependency from hotSwapPHPRuntime, where
`requestHandler.documentRoot` was accessed, and replaces it with an
explicit `cwd` argument. In the future, that argument will be removed
and the PHP constructor will accept a class-level `cwd` property.

 ## Testing instructions

Confirm the CI checks pass.