PHP: Remove setSapiName, setPhpIniEntry, setPhpIniPath methods from the remote PHP API client
setSapiName, setPhpIniEntry, setPhpIniPath are methods that
must either be called synchronously before the PHP internal
state is initialized, or with a complex argument that can't
be serialized over a remote connection.

They can only be used in the same process before PHP is
fully initialized and have never worked correctly when
connecting a Playground API client to `remote.html`.

Removing them makes that apparent, prevents confusing
API interactions.

In addition, it sets the stage for supporting
a [Loopback
Request](https://github.com/WordPress/wordpress-playground/pull/1287)
as the backend for the API endpoint switches from a single PHP instance
to a PHPRequestHandler, and here we're removing an API surface that
needs to impact all existing and future PHP instances.

 ## Testing instructions

Confirm the CI checks pass.