PHP.wasm Node: Revert a part of #1289, do not import a .wasm file (#1348)
https://github.com/WordPress/wordpress-playground/pull/1289 shipped a
PHP.wasm building logic that used an import statement instead of
composing a path of __dirname. It did that to support Bun bundling, as
Bun needs those static imports to include the .wasm files in the single
executable. Unfortunately, that change broke npm bundling and
publishing, resulting in a broken package that doesn't ship the correct
paths and builds a broken index.cjs file.