Website: Fix deploy-time check for file with PHP-handled redirect (#1350)
## What is this PR doing?

This PR solves an issue where `/wordpress.html` was not being served by
PHP because it was not moved into `files-to-serve-via-php` during
website deploy. And it was not being moved because the
should-serve-via-PHP check relied upon the current referer, which is
something we don't know at deploy time.

## How is the problem addressed?

This PR solves the issue by having the maybe-redirect function return a
declaration of its intent to redirect for specific referers. Then we can
see there is need for special treatment at deploy time, and the request
handler can see the declaration and act upon it at request time.

## Testing Instructions

- Tested manually via SSH on
playground-dot-wordpress-dot-net.atomicsites.blog.
- Will also test post-deploy by manually running the deploy workflow for
WP Cloud and confirming the fix afterward

Related to #1197