PHP: Support php.mv() between devices via recursive copy (#1411)
Before this PR, `php.mv()` from MEMFS to NODEFS would fail. With this
PR, it works.

Technically, this PR checks if `fromPath` and `toPath` belong to the
same mountpoint. If yes, it calls a fast `FS.rename()` function. If not,
it runs a slower recursive copy followed by recursive remove of the
`fromPath`.

This runs the entire recursive copy and only then runs the recursive
delete. I wonder if deleting each copied item as we go would make more
sense?

This PR also brings in the Emscripten TS types as they made working with
the `FS` object much easier.

## Testing Instructions

Review the new unit test and confirm it passes.