Remote.html: Preload the SQLite database plugin, but only execute it if there's no custom db.php inside wp-content (#1424)
This PR always preloads the SQLite plugin in remote.html, even if the
provided WordPress already ships one. The preloaded plugin will only be
used if no `wp-content/db.php` drop-in plugin is found in the
filesystem. This change sets the stage for:

* #1426, which ships a minified WordPress build without the SQLite
plugin baked-in.
* #1425, which runs the WordPress installation wizard if the site isn't
installed yet

Related to [Boot
Protocol](https://github.com/WordPress/wordpress-playground/issues/1398)

## Problem solved

We want the official WordPress release `.zip` to just work in
Playground. Currently it doesn't because remote.html expects a
pre-processed build with the SQLite plugin already baked-in. This PR, in
conjunction with #1425 and #1426, remove that requirement.

## Testing Instructions

Confirm the CI checks pass.