Remote.html: Install WordPress if it isn't installed yet (#1425)
## What is this PR doing?

Runs the WordPress installation process in remote.html when the loaded
WordPress instance isn't installed yet.

This is to enable the official WordPress release `.zip` to just work in
Playground. Currently it doesn't because remote.html expects a
pre-installed WordPress build with a pre-populated SQLite database
already baked-in. This PR, in conjunction with #1425 and #1426, remove
that requirement.

Related to [Boot
Protocol](https://github.com/WordPress/wordpress-playground/issues/1398)

## Testing Instructions

Confirm the CI checks pass.