PHP: Remove the MODE_EVAL_CODE execution mode
Removes the special MODE_EVAL_CODE in which the PHP code to evaluate
is fed to the interpreter as a string. Inline code snippets are now
written to a special file called `/internal/eval.php` and executed
as regular PHP files.

 ## Rationale

PHP does not correctly load the php.ini file when the code is passed as
a string.

This change also reduces the number of forks in the most cricitcal code
path in the project – PHP execution.

 ## Remaining work

* Rebuild all the PHP bundles, not just 8.0

 ## Testing instructions

Confirm the CI checks pass