Website: Make custom header conditions more specific
This is a general improvement, but it was done so we could
stop moving puzzle/index.html aside to be served by PHP.
When puzzle/index.html is set aside, it turns out that nginx
does not delegate the request to PHP but rather responds
with a 403 error. We plan to talk with the host about fixing
this behavior, but in the meantime, we can avoid it here.