Website: Stop disabling cache for non-root index.html files (#1461)
## What is this PR doing?

This PR makes it so the only `index.html` file set aside to get custom
HTTP headers is `/index.html`.

## What problem is it solving?

It is a general improvement for us to be more specific about which file
paths get special treatment, but we are making the change now so we can
stop setting puzzle/index.html aside to be served by PHP. cc @bgrgicak

## How is the problem addressed?

When puzzle/index.html is set aside, it turns out that nginx does not
delegate the request to PHP but rather responds with a 403 error. Nginx
appears to do this for every existing, non-root directory. We plan to
talk with the host about fixing this behavior, but in the meantime, we
can avoid it here by excluding puzzle/index.html from receiving
"no-cache" headers.

## Testing Instructions

I have tested these changes manually via SSH on a staging site. After
merging, we will deploy the website and confirm all is well in
production.