Blueprints: Remove support for the "request" step
Removes the "request" step. It is not very useful – most wp-admin requests
require a nonce, which it has no access to. Furthermore, it cannot be
easily ported to non-browser Playground runtimes

 ## Technical implementation

Removes the "request" step from any Blueprints when executing them.

 ## Follow-up work

Remove the request() function from the Blueprints package once no other
step depends on it.

 ## Testing instructions

Confirm the CI checks are green