Move Filesystem methods to a static FSHelpers class (#1468)
## What is this PR doing?

Move the implementation of the Filesystem-related methods from the PHP
class to a static FSHelpers class.

This is a non-breaking change. The PHP class still has the same
interface, it just delegates all the calls to FSHelpers.

The following filesystem methods are affected:
   - readFileAsText
   - readFileAsBuffer 
   - writeFile
   - unlink
   - mv
   - rmdir
   - listFiles
   - isDir
   - fileExists
   - mkdir
   - copyRecursive

## Motivation

This sets the stage for OPFSMountHandle

We keep reimplementing similar logic in different places. This PR makes
operations like recursive remove usable outside of the PHP class
instance.

## Testing Instructions

confirm the ci checks pass